feat(web): create dedicated service for private sql
    ref: MANAGER-13071
