+++ packages/manager/apps/web/client/app/hosting/database/hosting-database.constants.js
@@ -1,5 +1,4 @@
 export const PRIVATE_SQL_PLAN_CODE = 'private-sql-512';
-export const PRIVATE_SQL_PLAN_CODE_PREFIX = 'private-sql';
 export const WEBHOSTING_PRODUCT_NAME = 'webHosting';
 
 export const FREE_HOSTING_OFFER = ['HOSTING_FREE_100_M', 'START_10_M'];
