+++ packages/manager/apps/web/client/app/hosting/database/hosting-database.service.js
@@ -3,12 +3,7 @@ import forEach from 'lodash/forEach';
 import snakeCase from 'lodash/snakeCase';
 import some from 'lodash/some';
 
-import groupBy from 'lodash/groupBy';
-import {
-  PRIVATE_SQL_PLAN_CODE,
-  PRIVATE_SQL_PLAN_CODE_PREFIX,
-} from './hosting-database.constants';
-import { DB_OFFERS } from './order/public/hosting-database-order-public.constants';
+import { PRIVATE_SQL_PLAN_CODE } from './hosting-database.constants';
 
 angular.module('services').service(
   'HostingDatabase',
@@ -22,7 +17,6 @@ angular.module('services').service(
       OvhHttp,
       Poller,
       OvhApiOrderCatalogPublic,
-      WucOrderCartService,
       iceberg,
     ) {
       this.$http = $http;
@@ -33,97 +27,7 @@ angular.module('services').service(
       this.OvhHttp = OvhHttp;
       this.Poller = Poller;
       this.OvhApiOrderCatalogPublic = OvhApiOrderCatalogPublic;
-      this.WucOrderCartService = WucOrderCartService;
       this.iceberg = iceberg;
-    }
-
-    static getPrivateSqlCategory(catalog) {
-      return catalog.addons.reduce((acc, current) => {
-        const { planCode } = current;
-        if (planCode.startsWith(PRIVATE_SQL_PLAN_CODE_PREFIX)) {
-          const dbOffer = {
-            ...current,
-            productSize: current.product.split('-')[2],
-            tracking: DB_OFFERS.PRIVATE.TRACKING,
-          };
-          const dbms = dbOffer.configurations
-            .find(({ name }) => name === 'engine')
-            .values.map((db) => {
-              const [dbName, dbVersion] = db.split('_');
-              return { db, dbName, dbVersion };
-            });
-          const groupedDbms = groupBy(dbms, 'dbName');
-          dbOffer.engines = Object.keys(groupedDbms).map((dbGroup) => ({
-            dbGroup,
-            versions: groupedDbms[dbGroup],
-          }));
-          acc.push(dbOffer);
-        }
-        return acc;
-      }, []);
-    }
-
-    buildPrivateSqlDbCategories(webCloudCatalog) {
-      const webCloudCategory = this.constructor.getPrivateSqlCategory(
-        webCloudCatalog,
-      );
-      // const db groups
-      const groupedCategories = {
-        [DB_OFFERS.PRIVATE.CATEGORY]: {
-          versions: webCloudCategory,
-          tracking: DB_OFFERS.PRIVATE.TRACKING,
-          productName: DB_OFFERS.PRIVATE.PRODUCT_NAME,
-        },
-      };
-
-      const dbCategories = Object.keys(groupedCategories).map((category) => {
-        const { versions, tracking, productName } = groupedCategories[category];
-
-        return {
-          category,
-          versions,
-          tracking,
-          selectVersion: versions[0],
-          selectEngine: null,
-          productName,
-        };
-      });
-
-      return dbCategories;
-    }
-
-    buildDbCategories(catalog, webCloudCatalog) {
-      const webCloudCategory = this.constructor
-        .getWebCloudCategory(webCloudCatalog)
-        .sort(this.constructor.dbOfferSort);
-
-      // const db groups
-      const groupedCategories = {
-        [DB_OFFERS.PRIVATE.CATEGORY]: {
-          versions: webCloudCategory,
-          tracking: DB_OFFERS.PRIVATE.TRACKING,
-          productName: DB_OFFERS.PRIVATE.PRODUCT_NAME,
-        },
-      };
-
-      return Object.keys(groupedCategories).map((category) => {
-        const { versions, tracking, productName } = groupedCategories[category];
-
-        return {
-          category,
-          versions,
-          tracking,
-          selectVersion: versions[0],
-          selectEngine: null,
-          productName,
-        };
-      });
-    }
-
-    getPrivateSqlCatalogForHosting(ovhSubsidiary) {
-      return this.$http
-        .get(`/order/catalog/public/webHosting?ovhSubsidiary=${ovhSubsidiary}`)
-        .then(({ data }) => data);
     }
 
     /**
