+++ packages/manager/apps/web/client/app/hosting/database/private-sql-activation/index.js
@@ -1,5 +1,6 @@
 import privateSqlActivation from './private-sql-activation.component';
 import privateSqlActivationRouting from './private-sql-activation.routing';
+import service from './private-sql-activation.service';
 
 const moduleName = 'ovhManagerHostingDatabasePrivateSqlActivation';
 
@@ -7,6 +8,7 @@ angular
   .module(moduleName, [])
   .component('hostingDatabasePrivateSqlActivation', privateSqlActivation)
   .config(privateSqlActivationRouting)
+  .service('PrivateSql', service)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
