+++ packages/manager/apps/web/client/app/hosting/database/private-sql-activation/private-sql-activation.component.js
@@ -5,14 +5,12 @@ export default {
   controller,
   template,
   bindings: {
-    catalog: '<',
     goToHosting: '<',
     me: '<',
     dbCategories: '<',
     privateSqlCatalog: '<',
     hosting: '<',
     versions: '<',
-    datacenter: '<?',
     onError: '<',
     onSuccess: '<',
   },
