+++ packages/manager/apps/web/client/app/hosting/database/private-sql-activation/private-sql-activation.constants.js
@@ -1,6 +1,8 @@
 export const PRESELECTED_DB_CATEGORY = 'private';
 export const WEBHOSTING_PRODUCT_NAME = 'webHosting';
+export const PRIVATE_SQL_PLAN_CODE_PREFIX = 'private-sql';
 export default {
   PRESELECTED_DB_CATEGORY,
   WEBHOSTING_PRODUCT_NAME,
+  PRIVATE_SQL_PLAN_CODE_PREFIX,
 };
